requirements.txt
fastapi
pydantic
uvicorn
playwright
logging
pandas
numpy
statsmodels
yfinance
APScheduler
huggingface-hub
transformers
torch
Pillow
fitz
pytesseract
opencv-python
streamlit
plotly
opencv-python 
pymongo 
pytesseract 
httpx 
pymupdf 